refactor: complete domain restructuring phase
