+++ internal/core/kernel/commands/workflow_archive.go
@@ -1,6 +1,9 @@
 package commands
 
-import core "github.com/compozy/compozy/internal/core"
+import (
+	core "github.com/compozy/compozy/internal/core"
+	"github.com/compozy/compozy/internal/core/model"
+)
 
 // WorkflowArchiveCommand archives completed workflow directories.
 type WorkflowArchiveCommand struct {
@@ -12,7 +15,7 @@ type WorkflowArchiveCommand struct {
 
 // WorkflowArchiveResult wraps the existing archive result contract.
 type WorkflowArchiveResult struct {
-	Result *core.ArchiveResult
+	Result *model.ArchiveResult
 }
 
 // WorkflowArchiveFromConfig translates the legacy core.Config shape into a typed archive command.
@@ -26,7 +29,7 @@ func WorkflowArchiveFromConfig(cfg core.Config) WorkflowArchiveCommand {
 
 // WorkflowArchiveFromArchiveConfig translates the direct archive config into a
 // typed archive command without caller-side field copying.
-func WorkflowArchiveFromArchiveConfig(cfg core.ArchiveConfig) WorkflowArchiveCommand {
+func WorkflowArchiveFromArchiveConfig(cfg model.ArchiveConfig) WorkflowArchiveCommand {
 	return WorkflowArchiveCommand{
 		WorkspaceRoot: cfg.WorkspaceRoot,
 		RootDir:       cfg.RootDir,
@@ -36,8 +39,8 @@ func WorkflowArchiveFromArchiveConfig(cfg core.ArchiveConfig) WorkflowArchiveCom
 }
 
 // CoreConfig converts the command into the existing archive configuration shape.
-func (c WorkflowArchiveCommand) CoreConfig() core.ArchiveConfig {
-	return core.ArchiveConfig{
+func (c WorkflowArchiveCommand) CoreConfig() model.ArchiveConfig {
+	return model.ArchiveConfig{
 		WorkspaceRoot: c.WorkspaceRoot,
 		RootDir:       c.RootDir,
 		Name:          c.Name,
