+++ internal/core/kernel/commands/workspace_migrate.go
@@ -1,6 +1,9 @@
 package commands
 
-import core "github.com/compozy/compozy/internal/core"
+import (
+	core "github.com/compozy/compozy/internal/core"
+	"github.com/compozy/compozy/internal/core/model"
+)
 
 // WorkspaceMigrateCommand migrates workflow artifacts under one workspace target.
 type WorkspaceMigrateCommand struct {
@@ -14,7 +17,7 @@ type WorkspaceMigrateCommand struct {
 
 // WorkspaceMigrateResult wraps the existing migration result contract.
 type WorkspaceMigrateResult struct {
-	Result *core.MigrationResult
+	Result *model.MigrationResult
 }
 
 // WorkspaceMigrateFromConfig translates the legacy core.Config shape into a typed migration command.
@@ -30,7 +33,7 @@ func WorkspaceMigrateFromConfig(cfg core.Config) WorkspaceMigrateCommand {
 
 // WorkspaceMigrateFromMigrationConfig translates the direct migration config into
 // a typed migration command without lossy field remapping at the caller.
-func WorkspaceMigrateFromMigrationConfig(cfg core.MigrationConfig) WorkspaceMigrateCommand {
+func WorkspaceMigrateFromMigrationConfig(cfg model.MigrationConfig) WorkspaceMigrateCommand {
 	return WorkspaceMigrateCommand{
 		WorkspaceRoot: cfg.WorkspaceRoot,
 		RootDir:       cfg.RootDir,
@@ -42,8 +45,8 @@ func WorkspaceMigrateFromMigrationConfig(cfg core.MigrationConfig) WorkspaceMigr
 }
 
 // CoreConfig converts the command into the existing migration configuration shape.
-func (c WorkspaceMigrateCommand) CoreConfig() core.MigrationConfig {
-	return core.MigrationConfig{
+func (c WorkspaceMigrateCommand) CoreConfig() model.MigrationConfig {
+	return model.MigrationConfig{
 		WorkspaceRoot: c.WorkspaceRoot,
 		RootDir:       c.RootDir,
 		Name:          c.Name,
