+++ internal/core/plan/journal.go
@@ -1,4 +1,4 @@
-package preputil
+package plan
 
 import (
 	"context"
