+++ internal/core/prompt/common.go
@@ -3,29 +3,15 @@ package prompt
 import (
 	"crypto/sha256"
 	"encoding/hex"
-	"errors"
 	"fmt"
 	"os"
 	"path/filepath"
-	"regexp"
 	"sort"
-	"strconv"
 	"strings"
 
-	"github.com/compozy/compozy/internal/core/frontmatter"
 	"github.com/compozy/compozy/internal/core/model"
-	"gopkg.in/yaml.v3"
-)
-
-var (
-	ErrLegacyTaskMetadata   = errors.New("legacy XML task metadata detected")
-	ErrV1TaskMetadata       = errors.New("v1 task front matter detected")
-	ErrLegacyReviewMetadata = errors.New("legacy XML review metadata detected")
-	legacyStatusHeadingRe   = regexp.MustCompile(`(?mi)^##\s*status:`)
-	legacyTaskStatusRe      = regexp.MustCompile(`(?m)^##\s*status:\s*(\w+)`)
-	legacyReviewStatusRe    = regexp.MustCompile(`(?mi)^##\s*status:\s*([a-z]+)\b`)
-	taskFileNumberRe        = regexp.MustCompile(`^task_(\d+)\.md$`)
-	issueFileNumberRe       = regexp.MustCompile(`^issue_(\d+)\.md$`)
+	"github.com/compozy/compozy/internal/core/reviews"
+	"github.com/compozy/compozy/internal/core/tasks"
 )
 
 type BatchParams struct {
@@ -54,134 +40,6 @@ func BuildSystemPromptAddendum(p BatchParams) string {
 	return buildPRDSystemPromptAddendum(p.Memory)
 }
 
-func ParseTaskFile(content string) (model.TaskEntry, error) {
-	var node yaml.Node
-	if _, err := frontmatter.Parse(content, &node); err != nil {
-		if LooksLikeLegacyTaskFile(content) {
-			return model.TaskEntry{}, ErrLegacyTaskMetadata
-		}
-		return model.TaskEntry{}, fmt.Errorf("parse task front matter: %w", err)
-	}
-	if hasTaskV1FrontMatterKeys(&node) {
-		return model.TaskEntry{}, ErrV1TaskMetadata
-	}
-
-	var meta model.TaskFileMeta
-	if err := node.Decode(&meta); err != nil {
-		return model.TaskEntry{}, fmt.Errorf("decode task front matter: %w", err)
-	}
-
-	task := model.TaskEntry{
-		Content:      content,
-		Status:       strings.TrimSpace(meta.Status),
-		Title:        strings.TrimSpace(meta.Title),
-		TaskType:     strings.TrimSpace(meta.TaskType),
-		Complexity:   strings.TrimSpace(meta.Complexity),
-		Dependencies: normalizeDependencies(meta.Dependencies),
-	}
-	if task.Status == "" {
-		return model.TaskEntry{}, errors.New("task front matter missing status")
-	}
-	return task, nil
-}
-
-func ParseLegacyTaskFile(content string) (model.TaskEntry, error) {
-	if !LooksLikeLegacyTaskFile(content) {
-		return model.TaskEntry{}, errors.New("legacy task metadata not found")
-	}
-
-	task := model.TaskEntry{Content: content}
-	if m := legacyTaskStatusRe.FindStringSubmatch(content); len(m) > 1 {
-		task.Status = strings.TrimSpace(m[1])
-	}
-
-	contextStart := strings.Index(content, "<task_context>")
-	contextEnd := strings.Index(content, "</task_context>")
-	if contextStart == -1 || contextEnd <= contextStart {
-		return model.TaskEntry{}, errors.New("task_context block not found")
-	}
-
-	contextBlock := content[contextStart : contextEnd+len("</task_context>")]
-	task.TaskType = extractXMLTag(contextBlock, "type")
-	task.Complexity = extractXMLTag(contextBlock, "complexity")
-	task.Dependencies = normalizeLegacyDependencies(extractXMLTag(contextBlock, "dependencies"))
-	if task.Status == "" {
-		return model.TaskEntry{}, errors.New("legacy task status not found")
-	}
-	return task, nil
-}
-
-func hasTaskV1FrontMatterKeys(node *yaml.Node) bool {
-	mapping := node
-	if node.Kind == yaml.DocumentNode {
-		if len(node.Content) != 1 {
-			return false
-		}
-		mapping = node.Content[0]
-	}
-	if mapping.Kind != yaml.MappingNode {
-		return false
-	}
-	for idx := 0; idx+1 < len(mapping.Content); idx += 2 {
-		keyNode := mapping.Content[idx]
-		if keyNode.Kind != yaml.ScalarNode {
-			continue
-		}
-		switch strings.ToLower(strings.TrimSpace(keyNode.Value)) {
-		case "domain", "scope":
-			return true
-		}
-	}
-	return false
-}
-
-func LooksLikeLegacyTaskFile(content string) bool {
-	return strings.Contains(content, "<task_context>") ||
-		legacyStatusHeadingRe.MatchString(content)
-}
-
-func ExtractLegacyTaskBody(content string) (string, error) {
-	if !LooksLikeLegacyTaskFile(content) {
-		return "", errors.New("legacy task metadata not found")
-	}
-
-	lines := strings.Split(content, "\n")
-	body := make([]string, 0, len(lines))
-	inContext := false
-	for _, line := range lines {
-		trimmed := strings.TrimSpace(line)
-		switch {
-		case legacyStatusHeadingRe.MatchString(line):
-			continue
-		case trimmed == "<task_context>":
-			inContext = true
-			continue
-		case trimmed == "</task_context>":
-			inContext = false
-			continue
-		case inContext:
-			continue
-		default:
-			body = append(body, line)
-		}
-	}
-
-	return strings.TrimLeft(strings.Join(body, "\n"), "\n"), nil
-}
-
-func IsTaskCompleted(task model.TaskEntry) bool {
-	status := strings.ToLower(task.Status)
-	return status == "completed" || status == "done" || status == "finished"
-}
-
-func ExtractTaskNumber(filename string) int {
-	return extractFileNumber(filename, taskFileNumberRe)
-}
-
-func ExtractIssueNumber(filename string) int {
-	return extractFileNumber(filename, issueFileNumberRe)
-}
-
 func FlattenAndSortIssues(groups map[string][]model.IssueEntry, mode model.ExecutionMode) []model.IssueEntry {
 	allIssues := make([]model.IssueEntry, 0)
 	for _, items := range groups {
@@ -190,8 +48,8 @@ func FlattenAndSortIssues(groups map[string][]model.IssueEntry, mode model.Execu
 
 	if mode == model.ExecutionModePRDTasks {
 		sort.SliceStable(allIssues, func(i, j int) bool {
-			numI := ExtractTaskNumber(allIssues[i].Name)
-			numJ := ExtractTaskNumber(allIssues[j].Name)
+			numI := tasks.ExtractTaskNumber(allIssues[i].Name)
+			numJ := tasks.ExtractTaskNumber(allIssues[j].Name)
 			if numI != numJ {
 				return numI < numJ
 			}
@@ -201,8 +59,8 @@ func FlattenAndSortIssues(groups map[string][]model.IssueEntry, mode model.Execu
 	}
 
 	sort.SliceStable(allIssues, func(i, j int) bool {
-		numI := ExtractIssueNumber(allIssues[i].Name)
-		numJ := ExtractIssueNumber(allIssues[j].Name)
+		numI := reviews.ExtractIssueNumber(allIssues[i].Name)
+		numJ := reviews.ExtractIssueNumber(allIssues[j].Name)
 		if numI != 0 && numJ != 0 && numI != numJ {
 			return numI < numJ
 		}
@@ -253,7 +111,7 @@ func batchIssueRange(batchIssues []model.IssueEntry) (int, int, bool) {
 	maxIssue := 0
 	hasIssueRange := false
 	for _, issue := range batchIssues {
-		issueNum := ExtractIssueNumber(issue.Name)
+		issueNum := reviews.ExtractIssueNumber(issue.Name)
 		if issueNum == 0 {
 			continue
 		}
@@ -271,105 +129,6 @@ func batchIssueRange(batchIssues []model.IssueEntry) (int, int, bool) {
 		}
 	}
 	return minIssue, maxIssue, hasIssueRange
-}
-
-func ParseReviewContext(content string) (model.ReviewContext, error) {
-	var meta model.ReviewFileMeta
-	if _, err := frontmatter.Parse(content, &meta); err != nil {
-		if LooksLikeLegacyReviewFile(content) {
-			return model.ReviewContext{}, ErrLegacyReviewMetadata
-		}
-		return model.ReviewContext{}, fmt.Errorf("parse review front matter: %w", err)
-	}
-
-	ctx := model.ReviewContext{
-		Status:      strings.ToLower(strings.TrimSpace(meta.Status)),
-		File:        strings.TrimSpace(meta.File),
-		Line:        meta.Line,
-		Severity:    strings.TrimSpace(meta.Severity),
-		Author:      strings.TrimSpace(meta.Author),
-		ProviderRef: strings.TrimSpace(meta.ProviderRef),
-	}
-	if ctx.Status == "" {
-		return model.ReviewContext{}, errors.New("review front matter missing status")
-	}
-	return ctx, nil
-}
-
-func ParseLegacyReviewContext(content string) (model.ReviewContext, error) {
-	if !LooksLikeLegacyReviewFile(content) {
-		return model.ReviewContext{}, errors.New("legacy review metadata not found")
-	}
-
-	ctx := model.ReviewContext{
-		Status:      strings.ToLower(strings.TrimSpace(extractLegacyStatus(content))),
-		File:        extractXMLTag(content, "file"),
-		Severity:    extractXMLTag(content, "severity"),
-		Author:      extractXMLTag(content, "author"),
-		ProviderRef: extractXMLTag(content, "provider_ref"),
-	}
-	lineValue := strings.TrimSpace(extractXMLTag(content, "line"))
-	if lineValue != "" {
-		lineNumber, err := strconv.Atoi(lineValue)
-		if err != nil {
-			return model.ReviewContext{}, fmt.Errorf("parse legacy review line: %w", err)
-		}
-		ctx.Line = lineNumber
-	}
-	if ctx.Status == "" {
-		return model.ReviewContext{}, errors.New("legacy review status not found")
-	}
-	return ctx, nil
-}
-
-func LooksLikeLegacyReviewFile(content string) bool {
-	return strings.Contains(content, "<review_context>") ||
-		legacyStatusHeadingRe.MatchString(content)
-}
-
-func ExtractLegacyReviewBody(content string) (string, error) {
-	if !LooksLikeLegacyReviewFile(content) {
-		return "", errors.New("legacy review metadata not found")
-	}
-
-	lines := strings.Split(content, "\n")
-	body := make([]string, 0, len(lines))
-	inContext := false
-	for _, line := range lines {
-		trimmed := strings.TrimSpace(line)
-		switch {
-		case legacyStatusHeadingRe.MatchString(line):
-			continue
-		case trimmed == "<review_context>":
-			inContext = true
-			continue
-		case trimmed == "</review_context>":
-			inContext = false
-			continue
-		case inContext:
-			continue
-		default:
-			body = append(body, line)
-		}
-	}
-
-	return strings.TrimLeft(strings.Join(body, "\n"), "\n"), nil
-}
-
-func ParseReviewStatus(content string) (string, error) {
-	ctx, err := ParseReviewContext(content)
-	if err != nil {
-		return "", err
-	}
-	return ctx.Status, nil
-}
-
-func IsReviewResolved(content string) (bool, error) {
-	status, err := ParseReviewStatus(content)
-	if err != nil {
-		return false, err
-	}
-	return status == "resolved", nil
 }
 
 func sortCodeFiles(batchGroups map[string][]model.IssueEntry) []string {
@@ -416,8 +175,8 @@ func buildBatchChecklist(p BatchParams) string {
 		allIssues = append(allIssues, items...)
 	}
 	sort.Slice(allIssues, func(i, j int) bool {
-		numI := ExtractIssueNumber(allIssues[i].Name)
-		numJ := ExtractIssueNumber(allIssues[j].Name)
+		numI := reviews.ExtractIssueNumber(allIssues[i].Name)
+		numJ := reviews.ExtractIssueNumber(allIssues[j].Name)
 		if numI != 0 && numJ != 0 && numI != numJ {
 			return numI < numJ
 		}
@@ -440,32 +199,6 @@ func buildBatchChecklist(p BatchParams) string {
 	return chk.String()
 }
 
-func extractFileNumber(filename string, pattern *regexp.Regexp) int {
-	matches := pattern.FindStringSubmatch(filepath.Base(filename))
-	if len(matches) < 2 {
-		return 0
-	}
-	num, err := strconv.Atoi(matches[1])
-	if err != nil {
-		return 0
-	}
-	return num
-}
-
-func extractXMLTag(content, tag string) string {
-	openTag := "<" + tag + ">"
-	start := strings.Index(content, openTag)
-	if start < 0 {
-		return ""
-	}
-	start += len(openTag)
-	end := strings.Index(content[start:], "</"+tag+">")
-	if end < 0 {
-		return ""
-	}
-	return strings.TrimSpace(content[start : start+end])
-}
-
 func sanitizePath(path string) string {
 	runes := make([]rune, 0, len(path))
 	for _, r := range path {
@@ -477,37 +210,4 @@ func sanitizePath(path string) string {
 		runes = append(runes, '_')
 	}
 	return string(runes)
-}
-
-func normalizeDependencies(values []string) []string {
-	if len(values) == 0 {
-		return nil
-	}
-
-	normalized := make([]string, 0, len(values))
-	for _, value := range values {
-		trimmed := strings.TrimSpace(value)
-		if trimmed == "" || strings.EqualFold(trimmed, "none") {
-			continue
-		}
-		normalized = append(normalized, trimmed)
-	}
-	if len(normalized) == 0 {
-		return nil
-	}
-	return normalized
-}
-
-func normalizeLegacyDependencies(raw string) []string {
-	if strings.TrimSpace(raw) == "" {
-		return nil
-	}
-	return normalizeDependencies(strings.Split(raw, ","))
-}
-
-func extractLegacyStatus(content string) string {
-	if matches := legacyReviewStatusRe.FindStringSubmatch(content); len(matches) > 1 {
-		return matches[1]
-	}
-	return ""
 }
