+++ internal/core/workspace/config_validate.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/compozy/compozy/internal/core/agent"
 	"github.com/compozy/compozy/internal/core/model"
-	"github.com/compozy/compozy/internal/core/providers"
+	"github.com/compozy/compozy/internal/core/providerdefaults"
 	"github.com/compozy/compozy/internal/core/tasks"
 )
 
@@ -77,7 +77,7 @@ func validateFetchReviews(cfg FetchReviewsConfig) error {
 	if name == "" {
 		return errors.New("workspace config fetch_reviews.provider cannot be empty")
 	}
-	if _, err := providers.DefaultRegistry().Get(name); err != nil {
+	if _, err := providerdefaults.DefaultRegistry().Get(name); err != nil {
 		return fmt.Errorf("workspace config fetch_reviews.provider: %w", err)
 	}
 	return nil
